+++ CHANGELOG.md
@@ -6,7 +6,11 @@ A list of notable changes to the Simpl-Cloud library included in each release.
 1.0.9 (unreleased)
 ==================
 
-- Nothing changed yet.
+- Style updates for trigger buttons.
+
+- Wells update.
+
+- Player role design revisions.
 
 
 1.0.8 (26 April 2022)